[코드수정] DataBinding 준비, 기타 소스정리
