+++ app/src/main/java/com/example/kakaosearcher/kakaoaddress/model/AddressModel.kt
@@ -1,4 +1,4 @@
-package com.example.kakaosearcher.kakaoaddress.model.resmodel
+package com.example.kakaosearcher.kakaoaddress.model
 
 import com.google.gson.annotations.SerializedName
