+++ app/src/main/java/com/example/kakaosearcher/kakaoaddress/presenter/AddressPresenter.kt
@@ -1,6 +1,6 @@
 package com.example.kakaosearcher.kakaoaddress.presenter
 
-import com.example.kakaosearcher.kakaoaddress.model.resmodel.dto.AddressDto
+import com.example.kakaosearcher.kakaoaddress.model.dto.AddressDto
 import com.example.kakaosearcher.kakaoaddress.repository.AddressRepository
 import com.example.kakaosearcher.network.retrofit.CallBackListener
 import io.reactivex.disposables.CompositeDisposable
