+++ app/src/main/res/layout/fragment_main.xml
@@ -12,22 +12,21 @@
         android:layout_height="50dp"
         android:background="@null"
         android:gravity="right|center_vertical"
-        android:paddingEnd="10dp"
+        android:hint="@string/fragment_main_input_address"
         android:paddingStart="10dp"
-        android:hint="주소를 입력해주세요."
+        android:paddingEnd="10dp"
         android:singleLine="true"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toLeftOf="@id/btn_fragment_main_search"
+        app:layout_constraintEnd_toStartOf="@id/btn_fragment_main_search"
+        app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
     <Button
         android:id="@+id/btn_fragment_main_search"
         android:layout_width="wrap_content"
         android:layout_height="0dp"
-        android:text="검색"
+        android:text="@string/fragment_main_search_button_text"
         app:layout_constraintBottom_toBottomOf="@id/et_fragment_main_input_text"
-        app:layout_constraintLeft_toRightOf="@id/et_fragment_main_input_text"
-        app:layout_constraintRight_toRightOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
     <androidx.recyclerview.widget.RecyclerView
@@ -37,6 +36,5 @@
         android:scrollbars="vertical"
         app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
         app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/et_fragment_main_input_text"
-        />
+        app:layout_constraintTop_toBottomOf="@id/et_fragment_main_input_text" />
 </androidx.constraintlayout.widget.ConstraintLayout>
